Replace sort by search
Instead of sorting the entire list, just search through the (now unordered) list for the cell that will divide next. Even for a list that is already almost completely sorted, the sort function still needs to verify this for the entire list. Therefore, a search will always be faster than a sort.