Fix n_vs_t not recording from the last division to the end of the simulation
If the simulation ends because the time is up (instead of because the number of cells became 0 or too large), then no number of dividing cells was drawn from the last division to the end of the simulation. In plots, this caused a missing horizontal line from the last division to the end of the figure.